refactor: move KoreanTime structure from main.rs to domain/time.rs
