+++ rook/src/domain/mod.rs
@@ -3,9 +3,11 @@ mod email_verification_service;
 mod new_subscriber;
 mod repository_url;
 mod subscriber_email;
+mod time;
 
 pub use email_verification::*;
 pub use email_verification_service::*;
 pub use new_subscriber::*;
 pub use repository_url::*;
 pub use subscriber_email::*;
+pub use time::*;
